removed old update data file
