+++ updateData.py
@@ -1,19 +1,34 @@
-import json
+import os
+import sys
+from dotenv import load_dotenv
 from datetime import datetime, date, timedelta
+from supabase import create_client, Client
 
 
-# Dev tool atm for me to update the json data
-# This will be replaced by the front end once I implement backend w db
-# and user auth
+load_dotenv()
+
+url = os.getenv("SUPABASE_URL")
+key = os.getenv("SUPABASE_KEY")
+uid = os.getenv("PERSONAL_UID")
+supabase: Client = create_client(url, key)
+
+# I'm going to update a single array, and then push those changes into the
+# records instead of modifying and reading the entire json file
+
 
 # use dictionary instead to reduce repetitiveness
 task_mapping = {
-    'c': 'check studied',
     'b': 'bunpro daily lessons',
     'w': 'wanikani daily lessons',
     'r': 'review session',
     'a': 'additional study',
     'x': 'reset score'
+}
+
+id_mapping = {
+    'w': 1,
+    'b': 2,
+    'r': 3,
 }
 
 task_options = "\n".join(
@@ -25,97 +40,165 @@ completed_task_msg = (
     f'{task_options}\n'
 )
 
-with open('./src/data.json') as jsonData:
-    data = json.load(jsonData)
-
 
 def main():
-    # load_dates()
-    target_day = datetime.today()
-    day = input('Today or (Y)esterday? ')
-
-    if (day.lower() == 'y'):
-        target_day = target_day - timedelta(days=1)
-    target_day = target_day.strftime('%Y-%m-%d')
-
+    # if args are passed then it's being called as a subprocess
+    completed_tasks = []
+    study_date = datetime.today()
+    day = input('Today or insert date YYYY-MM-DD\t')
+    print('day is', day)
+    if (len(day)):
+        study_date = datetime.strptime(day, "%Y-%m-%d")
+    print(study_date)
     tasks = input(completed_task_msg).split(' ')
-
-    for task in tasks:
-        handle_completed_task(task, target_day)
-
-    print_points(target_day)
-
-    with open('./src/data.json', 'w') as jsonData:
-        json.dump(data, jsonData)
-
-
-def handle_completed_task(task, target_day):
-    if task in task_mapping:
+    tasks_to_process = [task for task in tasks if task in task_mapping]
+    for task in tasks_to_process:
         print(task_mapping[task])
-
-        if task == 'a':  # Additional study requires extra input
-            material = input('What did you study? ')
-            study_time = input('How long did you study? ')
-            add_points(material, target_day, study_time)
+        if task == 'a':
+            handle_additional_study(study_date)
         elif task == 'x':
-            reset_points(target_day)
-        elif task == 'c':
-            pass    # material gets printed from main, this is just a bypass
+            clear_study_records(study_date)
         else:
-            add_points(task, target_day)
+            material_id = id_mapping[task]
+            completed_tasks.append(material_id)
+    len(completed_tasks) > 0 and pushChanges(study_date, completed_tasks)
+    print_completed_tasks(study_date)
+
+
+def handle_additional_study(study_date):
+    material_title = input('What did you study? ').lower()
+    study_time = input('How long did you study? ')
+    material_data = get_additional_study(material_title)
+    if (not material_data):
+        material_id = add_material(material_title)
     else:
-        print('Invalid input')
+        material_id = material_data[0]['id']
+    insertStudiedItem(study_date, material_id)
+    updateStudiedItem(study_date, study_time, material_id)
 
 
-def add_date(given_date):
-    data.append({'date': given_date, 'studied': [], 'additional': {}})
+def add_material(material_title):
+    insert_response = (
+        supabase.table("study_material")
+        .insert({'user_id': uid, 'title': material_title})
+        .execute()
+    )
+    print('inserted new material', f"'{material_title}'", insert_response.data)
+    return insert_response.data[0]['id']
 
 
-def load_dates():
-    # tool to quickly load all dates into calendar
-    sdate = date(2025, 1, 1)
-    edate = date(2025, 12, 31)
-    currdate = sdate
-
-    while currdate <= edate:
-        curr_date_str = currdate.strftime('%Y-%m-%d')
-        if not date_exists(curr_date_str):
-            add_date(curr_date_str)
-        currdate += timedelta(days=1)
-
-    data.sort(key=lambda r: r['date'])
+def get_additional_study(material_title):
+    check_response = (
+        supabase.table("study_material")
+        .select('*')
+        .eq("title", material_title)
+        .eq("user_id", uid)
+        .execute()
+    )
+    return check_response.data
 
 
-def date_exists(given_date):
-    for item in data:
-        if item['date'] == given_date:
-            return True
-
-    return False
-
-
-def add_points(material, target_day, hours=None):
-    for x in data:
-        if (x['date'] == target_day):
-            if (hours):
-                print(material, hours)
-                x['additional'][material] = hours
-            else:
-                x['studied'].append(material)
+def clear_study_records(study_date):
+    response = (
+        supabase.table("studied_items")
+        .delete()
+        .eq("study_date", study_date)
+        .eq("user_id", uid)
+        .execute()
+    )
+    print('deleting items studied on', study_date)
+    print(len(response.data), 'item/s deleted')
 
 
-def reset_points(given_date):
-    for x in data:
-        if (x['date'] == given_date):
-            x['studied'] = []
-            x['additional'] = {}
+def print_completed_tasks(study_date):
+    response = (
+        supabase.table("study_records")
+        .select("user_profile(user_name), study_material(title), studied_items(count)")
+        .eq("study_date", study_date)
+        .eq("user_id", uid)
+        .execute()
+    )
+    if (response.data):
+        data = response.data[0]
+        study_material = data['study_material']
+        count = data['studied_items']
+        user_name = data['user_profile']['user_name']
+        print(user_name, 'studied on', study_date)
+        for x, y in zip(study_material, count):
+            print(x['title'], 'x', y['count'])
+    else:
+        print('nothing studied on', study_date)
 
 
-def print_points(given_date):
-    for x in data:
-        if (x['date'] == given_date):
-            print(f"Studied on {x['date']}:", x['studied'])
-            print('Additional study:', x['additional'])
+def doesRecordExist(study_date):
+    response = (
+        supabase.table("study_records")
+        .select("*")
+        .eq("study_date", study_date)
+        .eq("user_id", uid)
+        .execute()
+    )
+    return bool(response.data)
+
+
+def insertStudyRecord(study_date):
+    insert_response = (
+        supabase.table("study_records")
+        .insert({'study_date': study_date, 'user_id': uid})
+        .execute()
+    )
+    print("Inserted new record:", insert_response.data)
+
+
+def getStudyRow(study_date, material_id):
+    check_response = (
+        supabase.table("studied_items")
+        .select("count")
+        .eq("study_date", study_date)
+        .eq("user_id", uid)
+        .eq("material_id", material_id)
+        .execute()
+    )
+    return check_response.data
+
+
+def updateStudiedItem(study_date, new_count, material_id):
+    update_response = (
+        supabase.table("studied_items")
+        .update({'count': new_count})
+        .eq("study_date", study_date)
+        .eq("user_id", uid)
+        .eq("material_id", material_id)
+        .execute()
+    )
+    print('updated row:', update_response.data)
+
+
+def insertStudiedItem(study_date, material_id):
+    insert_response = (
+        supabase.table("studied_items")
+        .insert({'user_id': uid, 'study_date': study_date, 'material_id': material_id})
+        .execute()
+    )
+    print('inserted new row:', insert_response.data)
+
+
+def pushChanges(study_date, completed_tasks):
+    if doesRecordExist(study_date):
+        print('user with that record exists, skipping')
+    else:
+        insertStudyRecord(study_date)
+    for material_id in completed_tasks:
+        if material_id is None:
+            continue
+        rowData = getStudyRow(study_date, material_id)
+        if (rowData):
+            print('row already exists, updating count to: ',
+                  rowData[0]['count'])
+            new_count = int(rowData[0]['count']) + 1
+            updateStudiedItem(study_date, new_count, material_id)
+        else:
+            insertStudiedItem(study_date, material_id)
 
 
 main()
